+++ unstructured/partition/pdf_image/ocr.py
@@ -63,12 +63,16 @@ def process_data_with_ocr(
     Returns:
         DocumentLayout: The merged layout information obtained after OCR processing.
     """
-    with tempfile.NamedTemporaryFile() as tmp_file:
+    file_name = ""
+    with tempfile.NamedTemporaryFile(delete=False) as tmp_file:
         data_bytes = data if isinstance(data, bytes) else data.read()
         tmp_file.write(data_bytes)
         tmp_file.flush()
+        file_name = tmp_file.name
+
+    try:
         merged_layouts = process_file_with_ocr(
-            filename=tmp_file.name,
+            filename=file_name,
             out_layout=out_layout,
             extracted_layout=extracted_layout,
             is_image=is_image,
@@ -77,7 +81,11 @@ def process_data_with_ocr(
             ocr_mode=ocr_mode,
             pdf_image_dpi=pdf_image_dpi,
         )
-        return merged_layouts
+    finally:
+        if os.path.isfile(file_name):
+            os.remove(file_name)
+
+    return merged_layouts
 
 
 @requires_dependencies("unstructured_inference")